refactor(transfer): split progress, handshake, and writer
